+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Not yet released.
 * Dropped support for Python 3.6.
+* Improved detection of PWG 5100.13 based strings files.
 
 2.12
 ----